Remove m_passwordUpdate member variable
The m_passwordUpdate and m_tmpPassword variables' meaning is not clear.
It seems that m_passwordUpdate is useless at the moment, because
m_tmpPassword is the variable used to update the password too.

So, remove the m_passwordUpdate member; if we'll find a reason to have
another password variable, next time we'll document it better.