signond: immediate removal of SignonDisposables
SignonDisposable objects are removed from the disposableObjects list
when their destructor is run; since they are deleted with deleteLater(),
this means that they remain in the list for longer than necessary.
This is not a problem, but removing them from the list as soon as we
don't need them helps in keeping the list clean.