Inform signon-ui of identity removals and signouts
If the signon-ui implementation is storing some data related to the
identity (such as browser cookies), it might want to delete it when the
identity is deleted or is signing out.
Since this is a new D-Bus method that not all signon-ui implementations
might implement, don't panic if the D-Bus call fails (in any case, there
isn't much we could do about it).