Signond: reduce usage of SIGNOND_BUS macro
This macro is defined to be QDBusConnection::sessionBus(). Instead of
using it, let's use the connection stored in the QDBusContext,
currently, this leads to the same result (session bus), but this change
will make it easier to switch to p2p connections later on.

Also, fix a possible issue with messing up responses when different
clients call methods on the same identity at roughly the same time.