Signond: handle D-Bus registration in daemon class
Move the registration of D-Bus objects (Identity and AuthSession) to the
SignonDaemon class. This allows removing the hardcoded reference to the
session bus in Identity and AuthSession and will make it easier to
implement p2p support in the future.

Also, simplify code by not emitting user errors for OOM conditions.