lib: avoid race condition on destroyed remote objects
There's always been the following race condition in libsignon-qt, where
perfectly legitimate client calls could fail: that's in the small time
window between when signond has destroyed an unused object and the time
that this information has reached the client and has been processed by
libsignon-qt.

This commit fixes this issue, by making AsyncDBusProxy robust against
it: when we get the information that the remote object has been
destroyed, all the calls made on that object and not yet finished are
informed of the fact. Later on, when their D-Bus reply is received, it
is checked whether it's an error due to the object path having
disappeared and, if so, the call is re-scheduled to be executed once the
new object path is retrieved.
This relies on the fact that D-Bus preserves the ordering of the
messages; that is, the "unregistered" signal will always be received
before the error replies for the calls made after the object path was
unregistered.

In order to do this fix, the API of AsyncDBusProxy had to change
slightly, so there are a few adjustments to other parts of the code as
well.

Finally, the QTest::qWait() that was introduced to help avoid this race
condition could finally be removed.