signond: fix a memory leak and crash at exit
SignonIdentity was not destroying the SignonIdentityInfo in its
constructor.

Additionally, a crash was happening on the Qt4 build in the
SignonIdentity destructor, because it was accessing the
m_storedIdentities map of a half-destructed SignonDaemon object. Now the
logic has been reworked a bit to let the SignonDaemon monitor its
Identities, and to avoid the unnecessary exposure the SignonDaemon
object to the Identity class.

Last, the tracking of the unregistered identities in signond was
totally useless, so it has been removed.