signond: unregister the identity on disposable timeout
Currently, there's a small time window between the moment signond
removes all of its references to an Identity and the moment when the
identity is actually removed from D-Bus: this is caused by the call to
deleteLater().
If a client attempts to use an identity during this time window, the
call will be correctly be dispatched to the Identity, but the identity
will be deleted immediately afterwards.

We fix it by emitting the "destroyed" signal as soon as we decide that
the identity is going to be deleted. QDBusConnection will unregister the
Identity object from the bus when it gets this signal. The
AuthSessionCore class already uses this trick, so nothing needs to be
changed there.